ULK-332 5.1.0

+++ src/main/java/com/ulegalize/dto/IDossierDTO.java
@@ -141,5 +141,5 @@ public interface IDossierDTO {
      *
      * @return Cout horaire
      */
-    Integer getCouthoraire();
+    BigDecimal getCouthoraire();
 }
